feat: Implemented monitor writing to the database
The monitor is now able to log focus loss events in the database, using the DBHandler module to do so.

The DBHandler is the module that has exclusive access to database operations, which simplifies the architecture by having only one centralized element with all of the database interaction logic.

Additionaly, this handler maintains the connection to the database alive during the whole application lifetime, which removes the overhead of opening and closing these connections.